+++ Assignments/Entities/Truck.cs
@@ -8,7 +8,7 @@ namespace Assignment.Entities
 {
     public class Truck : Vehicles
     {
-        public Truck(VehicleTypes type, TruckBrands brand, TruckModel model, EngineType engine, int year, int mileage, decimal price) : base(type, year, mileage, price)
+        public Truck(VehicleTypes type, TruckBrands brand, TruckModel model, EngineType engine, int year, int mileage, decimal price) : base((int)type, year, mileage, price)
         {
             Brand = brand;
             Models = model;
